feat: Add notistack dependency for displaying notifications
The package.json file was modified to include the "notistack" dependency with version "^3.0.1". This package is used for displaying notifications in the application.

The BooksTable component was also updated to fix a bug where the key prop was incorrectly set to "books._id" instead of "index". This bug caused issues with rendering the table rows correctly.

In the main.jsx file, the BrowserRouter component was wrapped with the SnackbarProvider component from the "notistack" package. This allows the application to display notifications using the Snackbar component.

In the CreateBooks and EditBook pages, the useSnackbar hook from the "notistack" package was imported and used to display success and error notifications when creating or updating a book.

The BookSingleCard component was updated to include a new BiShow icon that triggers a modal when clicked. The modal displays additional information about the book.

This commit adds the necessary changes to include the "notistack" dependency and improve the user experience with notifications and book details modal.